backend1/requirements.txt
accelerate==0.33.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp_socks==0.9.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
autocorrect==2.6.1
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
bottle==0.12.25
Brotli==1.1.0
browser-cookie3==0.19.1
cachetools==5.4.0
cairocffi==1.7.1
CairoSVG==2.7.1
catalogue==1.0.2
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
cloudpathlib==0.18.1
clr-loader==0.2.6
colorama==0.4.6
comm==0.2.2
confection==0.1.5
configobj==5.0.8
configparser==7.0.0
contourpy==1.2.1
cryptography==43.0.0
cssselect2==0.7.0
curl_cffi==0.7.1
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docx2pdf==0.1.8
docx2txt==0.8
duckduckgo_search==6.2.11
etelemetry==0.3.1
executing==2.0.1
fastapi==0.112.2
filelock==3.15.4
filetype==1.2.0
fitz==0.0.1.dev2
Flask==3.0.3
fonttools==4.53.1
frontend==0.0.3
frozenlist==1.4.1
fsspec==2024.6.1
g4f==0.3.2.4
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
gpt4all==2.8.2
greenlet==3.0.3
grpcio==1.65.4
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.7
iopath==0.1.10
ipykernel==6.29.5
ipython==8.26.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-experimental==0.0.64
langchain-text-splitters==0.2.2
langcodes==3.4.0
langsmith==0.1.99
language_data==1.2.0
layoutparser==0.3.4
llamaapi==0.1.36
loguru==0.7.2
looseversion==1.3.0
lxml==5.3.0
lz4==4.3.3
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
nomic==3.1.1
numpy==1.26.4
openai==1.42.0
opencv-python==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathlib==1.0.1
pdf2image==1.17.0
pdf2img==0.1.2
pdfminer.six==20231228
pdfplumber==0.11.3
pillow==10.4.0
plac==1.1.3
platformdirs==4.2.2
plum-dispatch==1.7.4
plyer==2.1.0
portalocker==2.10.1
preshed==3.0.9
primp==0.6.1
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
prov==2.0.1
proxy_tools==0.1.0
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py4j==0.10.9.7
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydot==3.0.1
PyExecJS==1.5.1
Pygments==2.18.0
PyJWT==2.9.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
pyrsistent==0.20.0
pyspark==3.5.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-socks==2.5.1
pythonnet==3.0.3
pytz==2024.1
pywebview==5.2
pywin32==306
pyxnat==1.6.2
PyYAML==6.0.2
pyzmq==26.1.0
rdflib==6.3.2
referencing==0.35.1
regex==2024.7.24
reportlab==4.2.2
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.4
scipy==1.14.0
seaborn==0.13.2
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
Spire.Doc==12.7.1
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.38.2
sympy==1.13.2
tenacity==8.5.0
tesseract==0.1.3
textblob==0.18.0.post0
thop==0.1.1.post2209072238
tinycss2==1.3.0
tokenizers==0.19.1
torch==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
traits==6.3.2
transformers==4.44.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.0.196
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.5
wasabi==0.10.1
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.0.4
win32-setctime==1.1.0
wrapt==1.16.0
yarl==1.9.4
